[bfops/regen-dlls]: migrate changes
