+++ .cargo/config.toml
@@ -5,6 +5,7 @@ rustflags = ["--cfg", "tokio_unstable"]
 bump-versions = "run -p upgrade-version --"
 llm = "run --package xtask-llm-benchmark --bin llm_benchmark --"
 ci = "run -p ci --"
+regen = "run -p regen --"
 smoketest = "ci smoketests --"
 smoketests = "smoketest"
 lint = "ci lint --"
