+++ tools/regen/Cargo.toml
@@ -1,0 +1,10 @@
+[package]
+name = "regen"
+version = "0.1.0"
+edition.workspace = true
+
+[dependencies]
+anyhow.workspace = true
+clap.workspace = true
+duct.workspace = true
+tempfile.workspace = true
